refactor(solver): split the lost-in-space solve into explicit stages
The 450-line `solve_at_fov` mixed pattern enumeration, hash probing, the
SVD/parity step, the measured-FOV rebuild, verification, the pre-gate,
refinement, the post-refinement re-verify and the multiple-comparison
correction in one function, and `track.rs` re-implemented the front half.
Split it along the stage boundaries:

- `preprocess.rs`: `preprocess()` (CRPIX subtract, undistort, drop
  non-finite → `Preprocessed { centroids, orig_indices }`), the brightness
  sort and unit-vector builders, and `CentroidVectors { pixel_scale,
  parity_flip, data }` — verification vectors now carry the scale they
  were built at.
- `pattern_search.rs`: `PatternSearch` owns the FOV sweep, cluster-buster
  thinning, breadth-first combinations, key-range enumeration, hash-chain
  walk, edge-ratio check, Wahba SVD, parity detection and the measured-FOV
  rebuild, and hands each candidate to a caller-supplied acceptance
  callback as a `Hypothesis`. `build_fov_sweep` / key enumeration and
  their tests move with it.
- `verify.rs`: `verify_attitude`, `find_centroid_matches`, `binomial_cdf`,
  `diagonal_factor`. `verify_attitude` takes `CentroidVectors`.
- `solve.rs` keeps the driver (`solve_from_centroids`), the LIS
  acceptance stage (`accept_lis_candidate`: verify → pre-gate → refine →
  re-verify → sequential correction) with its per-solve `LisContext`,
  and refinement/finalization. The PR #59 re-verify rebuild is now
  expressed as "rebuild when the hypothesis vectors' scale differs from
  the refined scale".

Crate-private; the public API is unchanged. Behavior is bit-identical: a
1500-field synthetic dump (plain, 20 spurious, FOV bias with sweep,
parity-flipped, hinted tracking, aberration, wide sweep) matches the
pre-refactor output exactly apart from two `prob` values that also differ
between two runs of the *same* code — database generation iterates a
HashSet when building the pattern table, so chain order and the
candidates-tested divisor are not deterministic across processes (noted
for a follow-up). Unit, integration, skyview and TESS suites pass;
profiler 38.7 µs/solve easy, 1.82 ms no-match (unchanged within noise).
